Renamed variables and add Input class
