+++ Console_Calculator/bin/Debug/netcoreapp3.1/Console_Calculator.runtimeconfig.json
@@ -1,0 +1,9 @@
+{
+  "runtimeOptions": {
+    "tfm": "netcoreapp3.1",
+    "framework": {
+      "name": "Microsoft.NETCore.App",
+      "version": "3.1.0"
+    }
+  }
+}
