+++ HomeworkCalc/Calculator.cs
@@ -4,20 +4,7 @@ namespace HomeworkCalc
 {
     public class Calculator
     {
-
-        public static int GetValue() => int.Parse(Console.ReadLine());
-
-        public static string GetOperator() => Console.ReadLine();
-        public static string RequestOperator()
-        {
-            Console.WriteLine("Now write operation...");
-            return GetOperator();
-        }
-        public static int RequestValue()
-        {
-            Console.WriteLine("Enter the value...");
-            return GetValue();
-        }
+  
         static int Sum(int first, int second) => first + second;
         static int Difference(int first, int second) => first - second;
         public static int Divide(int first, int second)
@@ -27,9 +14,9 @@ namespace HomeworkCalc
         }
         static int Mult(int first, int second) => first * second;
 
-        public static int Calculate(int first, int second, string @operator)
+        public static int Calculate(int first, int second, string action)
         {
-            switch (@operator)
+            switch (action)
             {
                 case "+": return Sum(first, second);
                 case "-": return Difference(first, second);
